requirements.txt
# Django Core
Django==5.2.8
djangorestframework==3.16.1

# Database (PostgreSQL)
psycopg2-binary==2.9.11

# CORS Headers
django-cors-headers==4.9.0

# Environment Variables
python-decouple==3.8

# Static Files
whitenoise==6.6.0

# Image Processing
Pillow==10.4.0

# Rich Text Editor (Optional)
# django-ckeditor==6.7.0

# Dependencies
asgiref==3.10.0
sqlparse==0.5.3